+++ src/App.jsx
@@ -6,7 +6,7 @@ import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
 import Button from 'react-bootstrap/Button'
 import Image from 'react-bootstrap/Image'
-import { Profile, Gallery } from './components/Gallery'
+import { Gallery } from './components/Gallery'
 import { TodoList1, TodoList2 } from './components/TodoList'
